backend/requirements.txt
fastapi==0.115.14
uvicorn==0.34.3
SQLAlchemy==2.0.41
asyncpg==0.30.0
alembic==1.16.2
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.5.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
python-multipart==0.0.20